setup simple unit test
