+++ .eslintignore
@@ -1,0 +1,1 @@
+client/libraries.js
