+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0
 - Create 3D world.
+- Setup testing and linting.
 
 ## 1.0.0
 - Multi-player chat.
