+++ client/game.js
@@ -1,0 +1,5 @@
+const version = require('../package.json').version;
+
+module.exports = {
+  version
+};
