+++ client/libraries.js
